Request update
Add a single file includes the InnerProfile, add a fail state when details not saved, remove some unnecessary statement and attributes.
